+++ src/interfaces.ts
@@ -10,6 +10,7 @@ import type {
 import type DucksView from "./DucksView";
 import type MatrixView from "./MatrixView";
 import type StatsView from "./StatsView";
+import type {IJugglSettings} from "juggl-api";
 
 export type DebugLevel = keyof LogLevel;
 export interface BCSettings {
@@ -80,6 +81,7 @@ export interface BCSettings {
   showAll: boolean;
   showGrid: boolean;
   showImpliedRelations: boolean;
+  showJuggl: boolean;
   showPrevNext: boolean;
   showRefreshNotice: boolean;
   showTrail: boolean;
@@ -296,7 +298,7 @@ declare module "obsidian" {
   }
 }
 
-export interface ParsedCodeblock {
+export interface ParsedCodeblock extends IJugglSettings{
   dir: Directions;
   fields: string[];
   title: string;